backend/requirements.txt
absl-py==2.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.4.0
asttokens==3.0.1
astunparse==1.6.3
async-lru==2.0.5
attrs==25.4.0
babel==2.17.0
beautifulsoup4==4.14.2
bleach==6.3.0
blinker==1.9.0
cachetools==6.2.4
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
comm==0.2.3
contourpy==1.3.2
cryptography==45.0.5
cycler==0.12.1
debugpy==1.8.19
decorator==5.2.1
defusedxml==0.7.1
docutils==0.21.2
dotenv==0.9.9
executing==2.2.1
fastapi==0.128.0
fastjsonschema==2.21.2
filelock==3.20.0
filetype==1.2.0
Flask==3.1.2
flask-cors==6.0.2
flatbuffers==25.2.10
fonttools==4.59.0
fqdn==1.5.1
fsspec==2025.12.0
gast==0.6.0
gdown==5.2.0
git-filter-repo==2.47.0
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.45.0
google-auth-httplib2==0.3.0
google-generativeai==0.8.6
google-pasta==0.2.0
googleapis-common-protos==1.72.0
grpcio==1.73.1
grpcio-status==1.71.2
h11==0.16.0
h5py==3.14.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
huggingface-hub==0.36.0
idna==3.10
ipykernel==7.1.0
ipython==9.8.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.8
isoduration==20.11.0
itsdangerous==2.2.0
jax==0.7.1
jaxlib==0.7.1
jedi==0.19.2
Jinja2==3.1.6
jiwer==4.0.0
joblib==1.5.2
json5==0.12.1
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client==8.7.0
jupyter_core==5.9.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.3
jupyterlab==4.5.1
jupyterlab_pygments==0.3.0
jupyterlab_server==2.28.0
jupyterlab_widgets==3.0.16
keras==3.10.0
keras-tuner==1.4.7
Kivy==2.3.1
kivy-deps.angle==0.4.0
kivy-deps.glew==0.3.1
Kivy-Garden==0.1.5
kivy-garden.graph==0.4.0
kivy_deps.sdl2==0.8.0
kivy_garden.matplotlib==0.1.1.dev0
kivymd==1.2.0
kiwisolver==1.4.8
kt-legacy==1.0.5
lark==1.3.1
libclang==18.1.1
llvmlite==0.46.0
lxml==6.0.2
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.2.1
mdurl==0.1.2
mediapipe==0.10.21
mistune==3.2.0
ml_dtypes==0.5.1
more-itertools==10.8.0
mpmath==1.3.0
namex==0.1.0
nbclient==0.10.4
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.6.1
nltk==3.9.1
notebook==7.5.1
notebook_shim==0.2.4
numba==0.63.1
numpy==1.26.4
olefile==0.47
openai-whisper==20250625
opencv-contrib-python==4.11.0.86
opencv-python==4.9.0.80
opt_einsum==3.4.0
optree==0.16.0
packaging==25.0
pandas==2.3.1
pandocfilters==1.5.1
parso==0.8.5
pillow==11.3.0
platformdirs==4.5.1
plyer==2.1.0
portalocker==3.2.0
prometheus_client==0.23.1
prompt_toolkit==3.0.52
proto-plus==1.27.0
protobuf==4.25.8
psutil==7.2.1
pure_eval==0.2.3
pyarrow==22.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pyparsing==3.2.3
pypiwin32==223
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-json-logger==4.0.0
python-multipart==0.0.21
pytz==2025.2
pywin32==311
pywinpty==3.0.2
PyYAML==6.0.3
pyzmq==27.1.0
RapidFuzz==3.14.3
referencing==0.37.0
regex==2025.9.1
requests==2.32.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==14.0.0
rpds-py==0.30.0
rsa==4.9.1
sacrebleu==2.5.1
safetensors==0.7.0
scikit-learn==1.7.2
scipy==1.16.3
seaborn==0.13.2
Send2Trash==1.8.3
sentencepiece==0.2.1
setuptools==80.9.0
six==1.17.0
sounddevice==0.5.3
soupsieve==2.8
stack-data==0.6.3
starlette==0.50.0
sympy==1.14.0
tabulate==0.9.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
termcolor==3.1.0
terminado==0.18.1
threadpoolctl==3.6.0
tiktoken==0.12.0
tinycss2==1.4.0
tokenizers==0.22.2
torch==2.9.1
torchvision==0.24.1
tornado==6.5.4
tqdm==4.67.1
traitlets==5.14.3
transformers==4.57.3
typing-inspection==0.4.2
typing_extensions==4.14.1
tzdata==2025.2
uri-template==1.3.0
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.40.0
wcwidth==0.2.14
webcolors==25.10.0
webencodings==0.5.1
websocket-client==1.9.0
websockets==15.0.1
Werkzeug==3.1.3
wheel==0.45.1
widgetsnbextension==4.0.15
wrapt==1.17.2
youtube-dl==2021.12.17